Add files via upload
Changed the humidifier control logic. Previously, we always kept it at LOW. When we needed to turn on or off, we just pull it up to HIGH and pull it down to LOW again. The problem was that the humidifier could be randomly triggered when we turned on the heater. This is now fixed by keeping the humidifier at HIGH and performing a LOW-HIGH when we want to turn it on or off.